requirements.txt
anthropic
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
httpx
# AIDO pipeline
google-auth-oauthlib
google-api-python-client
